requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.4.1
django-cors-headers==3.10.0
djangorestframework==3.15.2
drf-spectacular==0.21.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nltk==3.9.1
oauthlib==3.2.2
openai==1.61.1
psycopg2-binary>=2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
